feat: Add flutter.create memo template
This commit adds a new file `flutter.create` with the content `{"template":"app"}`. This file is used as a template for creating Flutter apps. The addition of this template will make it easier for developers to create new Flutter projects with the desired structure.

Note: Please remove any meta information such as issue references, tags, or author names before committing.