refactor: Update pubspec.yaml and widget_test.dart
The `pubspec.yaml` file has been modified to update the `percent_indicator` dependency to version 4.2.3. This update includes bug fixes and improvements to the progress indicator functionality.

The `widget_test.dart` file has been refactored to disable the debug mode banner by setting the `debugShowCheckedModeBanner` property to `false` in the `MemoApp` widget. This change improves the visual appearance of the app during testing.